+++ NEWS.md
@@ -1,3 +1,10 @@
+# statgenGWAS 1.0.8
+
+* The gData object now has a plot function allowing plotting of the genetic map.
+* A bug in gData summaries when only one trait is present in the phenotypic data is fixed.
+* A bug in runSingleTraitGwas is fixed now allowing for using traits that are in pheno as covariates in covar.
+* Sorting of chromosomes in map is now done correctly for numeric values stored as character.
+
 # statgenGWAS 1.0.7
 
 * No user visual changes.